Implemented requested changes.

+++ src/main/resources/init_db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE `manufacturer_db` /*!40100 DEFAULT CHARACTER SET utf8 */ /*!80016 DEFAULT ENCRYPTION='N' */;
 
 CREATE TABLE `manufacturers` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(45) DEFAULT NULL,
   `country` varchar(45) DEFAULT NULL,
   `is_deleted` tinyint DEFAULT '0',
